Fixed warning of conversion of size_t to int issues
size_t is used to get container size which we are doing with the vectors.  The warnig is because size_t can be larger than an int.  in our case, this is not an issue because the size of the array will not be greater than an int.

+++ ELMS/base_station/main.cpp
@@ -121,7 +121,7 @@ int main()
 					// added to the vector of mine vehicles
 					vehicle = new Vehicle();
 					//set the index to the size of the mineVehicles
-					index = b.getMineVehicles().size();
+					index = static_cast<int>(b.getMineVehicles().size());
 					// add the vehicle pointer to the vector
 					b.addToMineVehicles(vehicle);
 					//this function is going to have a mutex and lock within the input_data function
